Former-commit-id: 1b30fe8fc04b5b4ce102d99a7dc64546c8245f0a
Former-commit-id: 3723f9cafc8bbaf1ddeea0d026b8e69cfe8bf136

+++ src/animation/easingFunctions.h
@@ -7,164 +7,164 @@ namespace phi
     {
     public:
         // no easing, no acceleration
-        static float linear(float t) { return t; }
+        static double linear(double t) { return t; }
 
         // accelerating from zero velocity
-        static float easeInQuad(float t) { return t*t; }
+        static double easeInQuad(double t) { return t*t; }
 
         // decelerating to zero velocity
-        static float easeOutQuad(float t) { return t * (2.0f - t); }
+        static double easeOutQuad(double t) { return t * (2.0 - t); }
 
         // acceleration until halfway, then deceleration
-        static float easeInOutQuad(float t) 
+        static double easeInOutQuad(double t) 
         { 
-            return t < 0.5f ? 
-                2.0f * t*t : 
-                -1.0f + (4.0f - 2.0f * t) * t; 
+            return t < 0.5 ? 
+                2.0 * t*t : 
+                -1.0 + (4.0 - 2.0 * t) * t; 
         }
 
         // accelerating from zero velocity
-        static float easeInCubic(float t) { return t*t*t; }
+        static double easeInCubic(double t) { return t*t*t; }
 
         // decelerating to zero velocity
-        static float easeOutCubic(float t) 
+        static double easeOutCubic(double t) 
         { 
             --t;
-            return t*t*t + 1.0f; 
+            return t*t*t + 1.0; 
         }
 
         // acceleration until halfway, then deceleration
-        static float easeInOutCubic(float t) 
+        static double easeInOutCubic(double t) 
         { 
-            return t < 0.5f ? 
-                4.0f * t*t*t : 
-                (t - 1.0f) * (2.0f * t - 2.0f) * (2.0f * t - 2.0f) + 1.0f; 
+            return t < 0.5 ? 
+                4.0 * t*t*t : 
+                (t - 1.0) * (2.0 * t - 2.0) * (2.0 * t - 2.0) + 1.0; 
         }
 
         // accelerating from zero velocity 
-        static float easeInQuart(float t) { return t*t*t*t; }
+        static double easeInQuart(double t) { return t*t*t*t; }
 
         // decelerating to zero velocity 
-        static float easeOutQuart(float t) 
+        static double easeOutQuart(double t) 
         {
             --t;
-            return 1.0f - t*t*t*t; 
+            return 1.0 - t*t*t*t; 
         }
 
         // acceleration until halfway, then deceleration
-        static float easeInOutQuart(float t) 
+        static double easeInOutQuart(double t) 
         {
-            if (t < 0.5f)
+            if (t < 0.5)
             {
-                return 8.0f * t*t*t*t;
+                return 8.0 * t*t*t*t;
             }
             else
             {
                 --t;
-                return 1.0f - 8.0f * t*t*t*t;
+                return 1.0 - 8.0 * t*t*t*t;
             }
         }
 
         // accelerating from zero velocity
-        static float easeInQuint(float t) { return t*t*t*t*t; }
+        static double easeInQuint(double t) { return t*t*t*t*t; }
 
         // decelerating to zero velocity
-        static float easeOutQuint(float t)
+        static double easeOutQuint(double t)
         { 
             --t;
-            return 1.0f + t*t*t*t*t; 
+            return 1.0 + t*t*t*t*t; 
         }
 
         // acceleration until halfway, then deceleration 
-        static float easeInOutQuint(float t)
+        static double easeInOutQuint(double t)
         {
-            if (t < 0.5f)
+            if (t < 0.5)
             {
-                return 16.0f * t*t*t*t*t;
+                return 16.0 * t*t*t*t*t;
             }
             else
             {
                 --t;
-                return 1.0f + 16.0f * t*t*t*t*t;
+                return 1.0 + 16.0 * t*t*t*t*t;
             }
         }
 
-        static float easeOutCirc(float t)
+        static double easeOutCirc(double t)
         {
             --t;
-            return glm::sqrt(1.0f - t*t);
+            return glm::sqrt(1.0 - t*t);
         };
 
-        static float easeOutBounce(float t)
+        static double easeOutBounce(double t)
         {
-            if (t < (1.0f / 2.75f))
+            if (t < (1.0 / 2.75))
             {
-                return 7.5625f * t * t;
+                return 7.5625 * t * t;
             }
-            else if (t < (2.0f / 2.75f))
+            else if (t < (2.0 / 2.75))
             {
-                t -= (1.5f / 2.75f);
-                return 7.5625f * t * t + 0.75f;
+                t -= (1.5 / 2.75);
+                return 7.5625 * t * t + 0.75;
             }
-            else if (t < (2.5f / 2.75f))
+            else if (t < (2.5 / 2.75))
             {
-                t -= (2.25f / 2.75f);
-                return 7.5625f * t * t + 0.9375f;
+                t -= (2.25 / 2.75);
+                return 7.5625 * t * t + 0.9375;
             }
             else
             {
-                t -= (2.625f / 2.75f);
-                return 7.5625f * t * t + 0.984375f;
+                t -= (2.625 / 2.75);
+                return 7.5625 * t * t + 0.984375;
             }
         }
 
-        static float easeOutElastic(float t)
+        static double easeOutElastic(double t)
         {
-            float s = 1.70158f;
-            float p = 0.0f;
-            float a = 1.0f;
+            double s = 1.70158;
+            double p = 0.0;
+            double a = 1.0;
             
-            if (t == 0.0f)
-                return 0.0f;
+            if (t == 0.0)
+                return 0.0;
             
-            if (t == 1.0f)
-                return 1.0f;
+            if (t == 1.0)
+                return 1.0;
             
             if (!p)
                 p = 0.3f;
 
-            if (a < 1.0f)
+            if (a < 1.0)
             {
-                a = 1.0f;
-                s = p / 4.0f;
+                a = 1.0;
+                s = p / 4.0;
             }
             else
             {
-                s = p / (2.0f * phi::PI) * std::asin(1.0f / a);
+                s = p / (2.0 * phi::PI) * std::asin(1.0 / a);
             }
 
-            return a * std::pow(2.0f, -10.0f * t) * std::sin((t - s) * (2.0f * phi::PI) / p) + 1.0f;
+            return a * std::pow(2.0, -10.0 * t) * std::sin((t - s) * (2.0 * phi::PI) / p) + 1.0;
         }
 
-        static float easeOutBack(float t, float s)
+        static double easeOutBack(double t, double s)
         {
             --t;
-            return t * t * ((s + 1.0f) * t + s) + 1.0f;
+            return t * t * ((s + 1.0) * t + s) + 1.0;
         }
 
-        static float easeOutBackDefault(float t)
+        static double easeOutBackDefault(double t)
         {
-            return easeOutBack(t, 1.70158f);
+            return easeOutBack(t, 1.70158);
         }
 
-        static float easeBounceBack(float t)
+        static double easeBounceBack(double t)
         {
-            return static_cast<float>(glm::sin(t * phi::PI));
+            return static_cast<double>(glm::sin(t * phi::PI));
         }
 
-        static float easeRubberBack(float t)
+        static double easeRubberBack(double t)
         {
-            return static_cast<float>(glm::sin(glm::pow(t, 1.0f / 2.0f) * phi::PI));
+            return static_cast<double>(glm::sin(glm::pow(t, 1.0 / 2.0) * phi::PI));
         }
     };
 }
